refactor: rename rule
